Remove --check-stop-position and --check-roundabouts for Winnipeg
PTNA has two consistent errors in its analysis of Winnipeg Transit:

1. PTv2 route: there is no 'public_transport' = 'stop_position'

Since Winnipeg Transit is just buses, the stop position is equivalent to the location of the pole, marked as public_transport=platform. We should disable the PTNA check.

2. PTv2 route: includes 1 entire roundabout but uses only segments

I think that we can remove this one, but I'm happy to be convinced otherwise by an expert. I see on the OSM Wiki that either including an entire roundabout or including segments is acceptable. The renderer will solve it.

I like having the roundabout be a complete way. In my experience, it has been easier to edit existing roundabouts when they are in a single way. But if there's a good reason why it should be split apart, I'm happy to hear it.